fix code

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,14 +6,14 @@
  */
 int main(void)
 {
-	unsigned int i, num1 = 0, num2 = 1, nextNum;
+	long unsigned int i, num1 = 0, num2 = 1, nextNum;
 
 	for (i = 0; i < 98; i++)
 	{
 		nextNum = num1 + num2;
 		num1 = num2;
 		num2 = nextNum;
-		printf("%u", nextNum);
+		printf("%lu", nextNum);
 		if (i != 97)
 			printf(", ");
 	}
